Fix for singleton / 65C02 issue
